third_party/blink/web_tests/platform/generic/fast/dom/Window/window-screen-properties-expected.txt

PASS: window.screen.pixelDepth is either 16, 24, or 32
PASS: window.screen.colorDepth is either 16, 24, or 32
